requirements.txt
coverage
docker>=3.5
gitpython
pykaniko
pymongo==3.7
pyyaml
semantic_version
wheel